+++ NEWS.md
@@ -1,5 +1,7 @@
 Version 0.7.7
  - change: "save" action will be enabled no matter contents in the current editor is modified or not
+ - fix: focus not correctly set when the current editor is closed
+ - fix: can't parse old c-style enum variable definition like "enum Test test;"
 
 Version 0.7.6
  - change: don't auto insert a new line when input an enter between '(' and ')' or between '[' and ']' (indent instead)